Thee_Kitchen/requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
python-dotenv==1.2.1
requests==2.31.0
Werkzeug==2.3.7
psycopg2-binary==2.9.7
